# docs/openapi.yaml
openapi: 3.0.1
info:
  title: LMS API
  version: 1.5.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetails:
      type: object
      description: Стандартный формат ошибок ASP.NET Core (RFC 7807)
      properties:
        type: { type: string, nullable: true }
        title: { type: string, nullable: true }
        status: { type: integer, nullable: true }
        detail: { type: string, nullable: true }
...
    AnswerBase:
      type: object
      required: [questionId, answerType]
      properties:
        questionId: { type: string, format: uuid }
        answerType: { type: string, enum: [SingleChoiceAnswer, MultipleChoiceAnswer, TextAnswer] }
    SingleChoiceAnswer:
      allOf:
        - $ref: '#/components/schemas/AnswerBase'
        - type: object
          required: [selectedOptionId]
          properties:
            selectedOptionId: { type: string, format: uuid }
    MultipleChoiceAnswer:
      allOf:
        - $ref: '#/components/schemas/AnswerBase'
        - type: object
          required: [selectedOptionIds]
          properties:
            selectedOptionIds:
              type: array
              items: { type: string, format: uuid }
              minItems: 1
    TextAnswer:
      allOf:
        - $ref: '#/components/schemas/AnswerBase'
        - type: object
          required: [text]
          properties:
            text: { type: string }
    AnswerItem:
      discriminator:
        propertyName: answerType
      oneOf:
        - $ref: '#/components/schemas/SingleChoiceAnswer'
        - $ref: '#/components/schemas/MultipleChoiceAnswer'
        - $ref: '#/components/schemas/TextAnswer'

    # ==========================================
    # РЕШЕНИЯ И ОЦЕНКИ
    # ==========================================
    Submission:
      type: object
      properties:
        id: { type: string, format: uuid }
        assignmentId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        answers:
          type: array
          items: { $ref: '#/components/schemas/AnswerItem' }
        status: { type: string, enum: [Draft, RequiresReview, Graded] }
        submittedAt: { type: string, format: date-time, nullable: true }
    SubmissionCreateRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items: { $ref: '#/components/schemas/AnswerItem' }
    Grade:
      type: object
      properties:
        id: { type: string, format: uuid }
        submissionId: { type: string, format: uuid }
        score: { type: integer }
        verdictText: { type: string }
        verdictedAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /assignments/{assignmentId}/submissions:
    get:
      tags: [Submissions]
      summary: Просмотр отправленных решений
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: isTeacher
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Submission' } } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Доступ запрещен.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    post:
      tags: [Submissions]
      summary: Создать решение (черновик)
      description: У созданного решения автоматически устанавливается статус Draft. Сервер обязан валидировать структуру ответов на соответствие типам вопросов (SingleChoice, MultipleChoice, Text).
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: isStudent
          in: query
          schema: { type: boolean, nullable: true }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmissionCreateRequest' }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        '400':
          description: Bad Request. Структура ответа не совпадает с типом вопроса (например, для Text вопроса прислан SingleChoiceAnswer).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Только STUDENT.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  /submissions/{submissionId}:
    get:
      tags: [Submissions]
      summary: Просмотр конкретного решения
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Только автор или TEACHER.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    patch:
      tags: [Submissions]
      summary: Редактировать контент решения
      description: Разрешено редактировать массив `answers` только если статус решения == Draft. Сервер выполняет строгую валидацию соответствия типа ответов типам вопросов.
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmissionCreateRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        '400':
          description: Bad Request. Ошибка валидации структуры ответов.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Только автор. Нельзя менять, если статус не Draft.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  /submissions/{submissionId}/submit:
    post:
      tags: [Submissions]
      summary: Отправить решение на проверку
      description: Смена статуса решения с Draft на RequiresReview. Передает решение преподавателю на проверку. Перед сменой статуса сервер может проверить, что все обязательные вопросы имеют ответ.
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Успешно отправлено
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        '400':
          description: Bad Request. Не на все вопросы дан ответ.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Только автор.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  /submissions/{submissionId}/withdraw:
    post:
      tags: [Submissions]
      summary: Отменить отправку решения
      description: Возвращает статус обратно с RequiresReview на Draft (возможно только если оценка еще не выставлена).
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Отправка отменена
          content: { application/json: { schema: { $ref: '#/components/schemas/Submission' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Только автор. Решение уже оценено.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }

  # ==========================================
  # GRADES
  # ==========================================
  /submissions/{submissionId}/grade:
    get:
      tags: [Grades]
      summary: Получить текущую оценку к решению
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Grade' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Только автор или TEACHER.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '404':
          description: Not Found. Оценка еще не выставлена.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    post:
      tags: [Grades]
      summary: Оценить решение
      description: Создает оценку и меняет статус Submission с RequiresReview на Graded.
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score: { type: integer }
                verdictText: { type: string }
      responses:
        '201':
          description: Оценка выставлена
          content: { application/json: { schema: { $ref: '#/components/schemas/Grade' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Оценить решение может любой пользователь с ролью ПРЕПОДАВАТЕЛЬ.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    put:
      tags: [Grades]
      summary: Обновить оценку решения
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score: { type: integer }
                verdictText: { type: string }
      responses:
        '200':
          description: Оценка обновлена
          content: { application/json: { schema: { $ref: '#/components/schemas/Grade' } } }
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Обновлять оценку может любой пользователь с ролью ПРЕПОДАВАТЕЛЬ.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '404':
          description: Not Found. Оценка еще не выставлена.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    delete:
      tags: [Grades]
      summary: Удалить оценку решения
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Оценка удалена
        '401':
          description: Unauthorized
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '403':
          description: Forbidden. Удалять оценку может любой пользователь с ролью ПРЕПОДАВАТЕЛЬ.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
        '404':
          description: Not Found. Оценка еще не выставлена.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }