# src/Backend/Api/Controllers/SubmissionsController.cs
using Api.Authentication;
using Application.Submissions.Contracts;
using Application.Submissions.Models;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

[ApiController]
[Route("api")]
public sealed class SubmissionsController : ControllerBase
{
    private readonly ISubmissionsService _submissionsService;

    public SubmissionsController(ISubmissionsService submissionsService)
    {
        _submissionsService = submissionsService;
    }
    [Authorize]

    [HttpPost("assignments/{assignmentId}/submissions")]
    public async Task<IActionResult> CreateSubmission(
        Guid assignmentId,
        [FromBody] SubmissionCreateRequest request,
        [FromQuery] bool? isStudent)
    {
        if (isStudent == false)
            return Forbid();

        var userId = User.GetUserId();
        if (userId is null)
            return Unauthorized();

        var result = await _submissionsService.CreateSubmission(assignmentId, userId.Value, request);

        if (result.Status == SubmissionAccessStatus.NotFound)
            return NotFound();

        if (result.Status == SubmissionAccessStatus.Forbidden)
            return Forbid();

        return Created($"/api/submissions/{result.Submission!.id}", result.Submission);
    }

    [HttpGet("assignments/{assignmentId}/submissions")]
    public async Task<IActionResult> GetSubmissions(
        Guid assignmentId,
        int limit = 20,
        int offset = 0,
        bool? isTeacher = true)
    {
        if (isTeacher == true)
        {
            var submissions = await _submissionsService.GetSubmissions(assignmentId, limit, offset);
            return Ok(submissions);
        }

        var userId = User.GetUserId();
        if (userId is null)
            return Unauthorized();

        var studentSubmissions = await _submissionsService.GetUserSubmissions(assignmentId, userId.Value, limit, offset);
        return Ok(studentSubmissions);
    }

    [HttpGet("submissions/{submissionId}")]
    public async Task<IActionResult> GetSubmission(Guid submissionId)
    {
        var result = await _submissionsService.GetSubmission(submissionId);

        if (result.Status == SubmissionAccessStatus.NotFound)
            return NotFound();

        if (result.Status == SubmissionAccessStatus.Forbidden)
            return Forbid();

        return Ok(result.Submission);
    }

    [HttpPatch("submissions/{submissionId}")]
    public async Task<IActionResult> PatchSubmission(
        Guid submissionId,
        [FromBody] SubmissionCreateRequest request)
    {
        var result = await _submissionsService.PatchSubmission(submissionId, request);

        if (result.Status == SubmissionAccessStatus.NotFound)
            return NotFound();

        if (result.Status == SubmissionAccessStatus.Forbidden)
            return Forbid();

        return Ok(result.Submission);
    }

    [HttpPost("submissions/{submissionId}/submit")]
    public async Task<IActionResult> SubmitSubmission(Guid submissionId)
    {
        var result = await _submissionsService.SubmitSubmission(submissionId);

        if (result.Status == SubmissionAccessStatus.NotFound)
            return NotFound();

        if (result.Status == SubmissionAccessStatus.Forbidden)
            return Forbid();

        return Ok(result.Submission);
    }

    [HttpPost("submissions/{submissionId}/withdraw")]
    public async Task<IActionResult> WithdrawSubmission(Guid submissionId)
    {
        var result = await _submissionsService.WithdrawSubmission(submissionId);

        if (result.Status == SubmissionAccessStatus.NotFound)
            return NotFound();

        if (result.Status == SubmissionAccessStatus.Forbidden)
            return Forbid();

        return Ok(result.Submission);
    }
}